requirements.txt
aiohttp
mutagen
loguru
python-dotenv